+++ 0x08-recursion/main.h
@@ -11,6 +11,6 @@ int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
 int is_prime_number(int n);
-int _sqrt_helper(int n, int low, int high);
+int _sqrt_helper(int n, long low, long high);
 
 #endif
